move startup log to platform directory
