+++ packages/dd-trace/src/platform/browser/startup-log.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = {
+  startupLog: () => {},
+  setStartupLogConfig: () => {},
+  setStartupLogPlugins: () => {},
+  setSamplingRules: () => {}
+}
